feature: support convert between two local DOM coord. And some refactor.
